CU-85zta35ra Create another router
